+++ product/src/main/java/com/owlexpress/product/domain/service/HubInfoServiceImpl.java
@@ -1,6 +1,6 @@
 package com.owlexpress.product.domain.service;
 
-import com.owlexpress.product.application.ProductHubUsecase;
+import com.owlexpress.product.application.ProductUsecase;
 import com.owlexpress.product.common.exceptions.ProductException;
 import com.owlexpress.product.domain.entity.HubInfo;
 import com.owlexpress.product.domain.repository.HubInfoRepository;
@@ -17,7 +17,7 @@ import java.util.UUID;
 public class HubInfoServiceImpl implements HubInfoService {
 
     private final HubInfoRepository hubInfoRepository;
-    private final ProductHubUsecase productHubUsecase;
+    private final ProductUsecase productUsecase;
 
     @Override
     @Transactional
@@ -48,7 +48,7 @@ public class HubInfoServiceImpl implements HubInfoService {
 
         hubInfo.updateModifiedData(1L);
         hubInfo.softDeleteData(1L);
-        productHubUsecase.disConnect(hubInfo);
+        productUsecase.disConnect(hubInfo);
     }
 
     private HubInfo getHubInfo(UUID hubInfoId) {
