+++ product/src/main/java/com/owlexpress/product/domain/service/ProductDomainService.java
@@ -1,8 +1,5 @@
 package com.owlexpress.product.domain.service;
 
-import java.util.UUID;
-
 public interface ProductDomainService {
 
-    void delete(UUID productsId);
 }
